requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
supabase==1.2.0
google-generativeai==0.5.0
praw==7.7.1
pytrends==4.9.2
Pillow==10.2.0
python-dotenv==1.0.1
schedule==1.2.1
lxml==5.1.0
httpx==0.24.1